Allow to create static machines for nodes that are not supposed to be reconciled
Instances can be created in advance before an actuator is deployed.
Some of the instances are not meant to be automatically reconciled.
Rather, managed by an external tool or a person.
Yet, it's still important to allow to create such machines so tests
assuming the same number of nodes as machines can pass.